requirements.txt
gunicorn==20.0.4
flask==1.1.1
flask-restful==0.3.8
faiss-cpu==1.6.1
numpy==1.18.1
happybase==1.2.0
json5==0.9.1
multiprocess==0.70.9
ujson==1.35
pyyaml==5.3
pymysql==0.9.3
six==1.12.0
auth==0.5.3
cacheout==0.11.2
schedule==0.6.0
kazoo==2.8.0
asyncio==3.4.3
grpcio==1.16.1
grpcio-tools==1.16.1